fix: 일정 반경 내 좌표를 가진 데이터 셋 추출 Query 변경 in StoreDao::getStoresByGeo

+++ app/src/main/java/com/minux/mask_alarmi/data/model/Store.kt
@@ -6,8 +6,8 @@ import androidx.room.PrimaryKey
 @Entity
 data class Store (
     @PrimaryKey val code: String,
-    val lat: Long,
-    val lng: Long,
+    val lat: Double,
+    val lng: Double,
     val addr: String,
     val name: String,
     val type: String,
